Simplifying API
Removed `is_toolarge` from `gretchenTX_inspect_t` struct and accordingly
from the implementation.
Since its a useful functioniality (checking if the filesize is too large) i added
it to the testtx.c file. It should be up to the implementor to deceide
what makes a too_big file. So i remove any assumption from the library
as such and leave it up to the user of the lib to deceide.
Also has the benefit of less code in the lib.