Gretchen
A bit of function renaming in modem(tx and rx).
And also this commit fixes the behavior of swallowing the last frame on the rx side.
Which was caused by the audio backend. Playing back the generated sample sometimes
resulted in a `click` sound at the end, which rendered the last frame undecodeable.
By adding half a second of silence to the generated sample (TX) the
problem is solved elegantly.
Note at the highest level of abstraction of gretchenRX there is no more `flush`
to extract the last (smaller) frame from the modem.