Gretchen
Decided to not fix the crash when using opt5 *.
But i identified the casue:
If the specified `framelen` parameter (in options) is too small for the corresponding
static values of `chunk_want` in `gretchenTX` (tx.c:84) then the audio
backend later will crash starting the stream.
Valgrind showed an `invalid write of 8` while generating the sample but
i could not identify the source.
So in all the tx-modem generates a sample which causes the audio backend
to crash if it is played back.

Some value combinations:
framelen    chunk size      -
> 300       1<10            ok
256         1<10            CRASH!
256         1<8             ok

This should be put into the developer documentation.

*
A good fix would be to check the sample before playing it back. This
seems to be a better approach since all possible bad combinations cannot
be inferred by me in advance.

But i do not plan to make the modem parameters accessible (which would
be counterproductive) and therefore writing lots of code for mitigating
problems which should never occur to a user an application kaing use of
the library makes no sense.