Gretchen CMakeLists.txt
removed the old code for finding the dnymaic dependency of libliquid.

NOTE to reduce an unused dependency in libliquid:
liliquid needs to be compile without using fftw3
    change config.h
    which is generated and overwritten by ./configure so do this after the configure step

    set
    #define HAVE_FFTW3_H 0

    and set
    #define HAVE_LIBFFTW3F 0

then compile (make) liquid and copy the libliquid.a and the headerfiles from include/ to th    e external/lib/liquid directoty.
The same compile and copy step needs to be done with libfec.