Record and Play arguments
Fixed old missconception of how to use POSIX getopt.
It turns out it will complain if an option without an argument
is passed e.g. just 'myprog -s'.
So in order to have binary options (like in python) i have to do
'myprog -s 1'. There seems to be a gnu extension of getopt that would
allow this. But this is of the plate i dont want to cause any more linking
hassles for no benefit (calling the programs from bash scripts).