Fixed the pa WTF recording issue
In short:
portaudio debug message spam resulted in me overlooking a getopt error
message `option requires an argument -- 's'` which resulted in an
important boolean flag to be always false (set to record mono)! This
further resulted in me interpreting the output of the mono-recording as
stereo and hence having the `portaudio WTF problem` of recording in the
wrong samplerate dispite following the examples 100%!
So my actual code was correct and the root cause was, as always, somewhere
else!

O M G

I love c :D
It is so 70s.
Its the time i was born.