Backend
Changed the initialize backend API to request a number of recording channels.
If the default device does not support that amount NULL is returned.
TODO
make the record callback make use of arbitrary number of channels.

Background:

It seems that it is not possible in portaudio to
query the exact amount of input channels...
https://stackoverflow.com/questions/40176632/how-to-discover-a-list-of-physical-audio-input-channels

I get on my arch box the value of maxInputChannels == 128!!!
that is a bit too much and i guess my box has not 128 microphones...
const PaDeviceInfo* info = Pa_GetDeviceInfo(back->strParams.device);
back->strParams.channelCount = info->maxInputChannels;