Backend
Added input channel estimation to the backend module. Also changed the
API again to not specify the number of desired channels.
For now, if in tx mode, the backend will request a mono channel to play
the sample on (we will see whether it will be better forcing to just use
one output channel and not n playing the same; TODO evaluate!)
If in rx mode the backend will estimate the number of real input
channels (e.g. stereo==2) and will discard everything but the first
channel to avoid smearing the received message. (Avoiding this has been a major
improvement. OMG)

The estimation is simple but slow:
start with 1 channel and increment the number after testing:
    open a stream an analyze the data in the callback
    if any channel contains only zeros we interpret this to be a virtual channel.
    if so we stop and report the current number -1