Modem
Added explicit methods to set the callback functions in the RX and TX structs.
In plain c unnessecarry but in swift normal c struct members are not accessible
without writing a wrapper.
Also helps the user of the lib to recognize the intent more easily; also keeps
my swift code simpler. (We will see about that)