MODEM TX
in tx.c `_prepare_for_transmission` added check if old sample data is still
present (and subsequently to remove it) before generating a new sample.

took me a while to find that the ios app was not the culprit in this
unwanted appending behavior.

NICE it shows that testing (or `real world` usage) helps.