Modem Helper
Fixing problems on osx with `envelope_unpack` where the binary
delimiter char does not produce meaningful results with strtok on osx
(truncating the whole string instead of splitting it).

Therefore i choose a normal (no binary) delimiter ";;". This mandates
to check the occurance of the delimiter string in the `envelope_create` method,
to mitigate splitting at the wrong position.